Load m1n1 at top of kernel data
m1n1 expects itself to be loaded below top of kernel data, so load it
right at top of kernel data and then update top of kernel so that it is
below top of kernel data.

In particular, its memory allocator start at topOfKernelData and assumes
all memory between that and physBase + memSize is free. More severely,
the chainloading scripts will derive new topOfKernelData from the
current m1n1, so the current approach of loading would actually cause
chainloading to set topOfKernelData to near the end of memory, causing
problems.m1n1 expects itself to be loaded below top of kernel data, so
do it.

Co-authored-by: Siguza <siguza@siguza.net>
Signed-off-by: Nick Chan <towinchenmi@gmail.com>